move to a separate package instead and address comments
