+++ packages/backend-common/package.json
@@ -41,7 +41,6 @@
     "@types/cors": "^2.8.6",
     "@types/dockerode": "^3.2.1",
     "@types/express": "^4.17.6",
-    "@types/luxon": "^2.0.4",
     "archiver": "^5.0.2",
     "aws-sdk": "^2.840.0",
     "compression": "^1.7.4",
@@ -60,7 +59,6 @@
     "knex": "^0.95.1",
     "lodash": "^4.17.21",
     "logform": "^2.1.1",
-    "luxon": "^2.0.2",
     "minimatch": "^3.0.4",
     "minimist": "^1.2.5",
     "morgan": "^1.10.0",
@@ -69,10 +67,8 @@
     "stoppable": "^1.1.0",
     "tar": "^6.1.2",
     "unzipper": "^0.10.11",
-    "uuid": "^8.0.0",
     "winston": "^3.2.1",
-    "yn": "^4.0.0",
-    "zod": "^3.9.5"
+    "yn": "^4.0.0"
   },
   "peerDependencies": {
     "pg-connection-string": "^2.3.0"
@@ -85,7 +81,6 @@
   "devDependencies": {
     "@backstage/backend-test-utils": "^0.1.8",
     "@backstage/cli": "^0.8.2",
-    "@backstage/test-utils": "^0.1.21",
     "@types/archiver": "^5.1.0",
     "@types/compression": "^1.7.0",
     "@types/concat-stream": "^1.6.0",
@@ -108,12 +103,10 @@
     "msw": "^0.35.0",
     "mysql2": "^2.2.5",
     "recursive-readdir": "^2.2.2",
-    "supertest": "^6.1.3",
-    "wait-for-expect": "^3.0.2"
+    "supertest": "^6.1.3"
   },
   "files": [
     "dist",
-    "migrations/**/*.{js,d.ts}",
     "config.d.ts"
   ],
   "configSchema": "config.d.ts"
