+++ packages/backend-tasks/api-report.md
@@ -1,0 +1,59 @@
+## API Report File for "@backstage/backend-tasks"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { Config } from '@backstage/config';
+import { DatabaseManager } from '@backstage/backend-common';
+import { Duration } from 'luxon';
+import { Logger as Logger_2 } from 'winston';
+
+// @public
+export interface LockOptions {
+  timeout: Duration;
+}
+
+// @public
+export interface PluginTaskManager {
+  acquireLock(
+    id: string,
+    options: LockOptions,
+  ): Promise<
+    | {
+        acquired: false;
+      }
+    | {
+        acquired: true;
+        release(): Promise<void>;
+      }
+  >;
+  scheduleTask(
+    id: string,
+    options: TaskOptions,
+    fn: () => void | Promise<void>,
+  ): Promise<{
+    unschedule: () => Promise<void>;
+  }>;
+}
+
+// @public
+export class TaskManager {
+  constructor(databaseManager: DatabaseManager, logger: Logger_2);
+  forPlugin(pluginId: string): PluginTaskManager;
+  // (undocumented)
+  static fromConfig(
+    config: Config,
+    options?: {
+      databaseManager?: DatabaseManager;
+      logger?: Logger_2;
+    },
+  ): TaskManager;
+}
+
+// @public
+export interface TaskOptions {
+  frequency?: Duration;
+  initialDelay?: Duration;
+  timeout?: Duration;
+}
+```
