+++ packages/backend-tasks/package.json
@@ -1,0 +1,54 @@
+{
+  "name": "@backstage/backend-tasks",
+  "description": "Common distributed task management / locking library for Backstage backends",
+  "version": "0.1.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "private": false,
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "types": "dist/index.d.ts"
+  },
+  "homepage": "https://backstage.io",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/backstage/backstage",
+    "directory": "packages/backend-tasks"
+  },
+  "keywords": [
+    "backstage"
+  ],
+  "license": "Apache-2.0",
+  "scripts": {
+    "build": "backstage-cli build --outputs cjs,types",
+    "lint": "backstage-cli lint",
+    "test": "backstage-cli test",
+    "prepack": "backstage-cli prepack",
+    "postpack": "backstage-cli postpack",
+    "clean": "backstage-cli clean"
+  },
+  "dependencies": {
+    "@backstage/backend-common": "^0.9.7",
+    "@backstage/config": "^0.1.10",
+    "@backstage/errors": "^0.1.3",
+    "@backstage/types": "^0.1.1",
+    "@types/luxon": "^2.0.4",
+    "knex": "^0.95.1",
+    "lodash": "^4.17.21",
+    "luxon": "^2.0.2",
+    "uuid": "^8.0.0",
+    "winston": "^3.2.1",
+    "zod": "^3.9.5"
+  },
+  "devDependencies": {
+    "@backstage/backend-test-utils": "^0.1.7",
+    "@backstage/cli": "^0.8.0",
+    "jest": "^26.0.1",
+    "wait-for-expect": "^3.0.2"
+  },
+  "files": [
+    "dist",
+    "migrations/**/*.{js,d.ts}"
+  ]
+}
