+++ packages/backend-tasks/src/setupTests.ts
@@ -14,21 +14,4 @@
  * limitations under the License.
  */
 
-/**
- * Common functionality library for Backstage backends
- *
- * @packageDocumentation
- */
-
-export * from './cache';
-export { loadBackendConfig } from './config';
-export * from './database';
-export * from './discovery';
-export * from './hot';
-export * from './logging';
-export * from './middleware';
-export * from './paths';
-export * from './reading';
-export * from './scm';
-export * from './service';
-export * from './util';
+export {};
